face-attendance-system/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
opencv-python==4.12.0.88
face-recognition==1.3.0
numpy==2.2.6
ultralytics==8.3.157
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
